refactor: move DirName to separate file; add peek() to KeyParser
