+++ src/meta/kvapi/src/kvapi/key_parser.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::iter::Peekable;
 use std::str::Split;
 use std::string::FromUtf8Error;
 
@@ -29,7 +30,7 @@ pub struct KeyParser<'s> {
     i: usize,
     /// The index of char for the next return.
     index: usize,
-    elements: Split<'s, char>,
+    elements: Peekable<Split<'s, char>>,
 }
 
 impl<'s> KeyParser<'s> {
@@ -39,20 +40,19 @@ impl<'s> KeyParser<'s> {
             str_key: source,
             i: 0,
             index: 0,
-            elements: source.split('/'),
+            elements: source.split('/').peekable(),
         }
     }
 
     /// Similar to `new()` but expect the specified prefix as the first part.
     pub fn new_prefixed(source: &'s str, prefix: &str) -> Result<Self, KeyError> {
-        let mut s = Self {
-            str_key: source,
-            i: 0,
-            index: 0,
-            elements: source.split('/'),
-        };
+        let mut s = Self::new(source);
         s.next_literal(prefix)?;
         Ok(s)
+    }
+
+    pub fn source(&self) -> &str {
+        self.str_key
     }
 
     /// Get the index of the last returned element.
@@ -60,6 +60,27 @@ impl<'s> KeyParser<'s> {
     /// If no element is returned, it will panic.
     pub fn index(&self) -> usize {
         self.i - 1
+    }
+
+    /// Peek the next element in raw `&str` and expect it to be `expect`, without unescaping or decoding.
+    pub fn peek_literal(&mut self, expect: &str) -> Result<(), KeyError> {
+        let next = self
+            .elements
+            .peek()
+            .ok_or_else(|| KeyError::WrongNumberOfSegments {
+                expect: self.i,
+                got: self.str_key.to_string(),
+            })?;
+
+        if *next != expect {
+            return Err(KeyError::InvalidSegment {
+                i: self.i,
+                expect: expect.to_string(),
+                got: next.to_string(),
+            });
+        }
+
+        Ok(())
     }
 
     /// Pop the next element in raw `&str`, without unescaping or decoding.
@@ -288,6 +309,21 @@ mod tests {
         assert!(kp.next_literal("bar%20-").is_ok());
         assert!(kp.next_literal("123").is_ok());
         assert!(kp.done().is_ok());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_key_parser_peek_literal() -> anyhow::Result<()> {
+        let s = "_foo/bar%20-/123";
+
+        let mut kp = KeyParser::new(s);
+        assert!(kp.peek_literal("_foo").is_ok());
+        assert_eq!(
+            kp.peek_literal("bar").unwrap_err().to_string(),
+            "Expect 0-th segment to be 'bar', but: '_foo'"
+        );
+        assert!(kp.next_literal("_foo").is_ok());
 
         Ok(())
     }
